Fix bug in live/servo.c
The servo returns a `uint16_t' but the value was being cast from an `int'.
This is fine for the return value but when it is sent along network
there is a size mismatch and it segfaults.

Changed the static `current_pos' value to type `uint16_t' to fix this.

+++ include/live/servo.c
@@ -3,7 +3,7 @@
 #include "servo.h"
 #include "network.h"
 
-static int current_pos = 0;
+static uint16_t current_pos = 0;
 
 void servo_init(){
     pwm_init();
